refactor:#27-Map관련 공통되는 부분 컴포넌트화 진행
